api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.4
python-dotenv==1.0.1
boto3==1.34.0
redis==5.0.1
opencv-python==4.8.1.78
numpy>=1.24.0
mediapipe>=0.10.8
celery==5.3.4
